+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+### Added
+- Per connection http provider configuration
+
+### Changed
+- deprecated K8s.http_provider/0 for K8s.default_http_provider/0
+
+## [0.5.2] - 2020-07-31
+
+### Added
+- Added auth `exec` support
+
+## [0.5.1] - 2020-07-17
 
 ### Added
 - K8s.Operation struct `query_params` field
@@ -14,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deprecated HTTPoison options being passed to K8s.Client.Runner.base
 - K8s.Operation.put_query_param/3 to add query parameters by key 
 - K8s.Operation.get_query_param/3 to get a query parameter by key
+- DigitalOcean authentication
+- `K8s.Resource.NamedList.access/1` - Accessor for lists with named items (containers, env, ...) ([#82](https://github.com/coryodaniel/k8s/pull/82))
 
 ### Changed
 - Refactored old references to `cluster_name` to `conn`